Add basic support for standard out logging
The support is limited to the Logstash JSON format as provided by the
logback extension.

When used, the Logspy extension intercepts System.out by rerouting it. The
interception happens transparently. Other well-behaving tools that
reroute System.out will work as before. The rerouting will only happen
during tests.

Noteworthy changes

The spied event now contains a copy of the thrown exception instead of the
original exception. Unlike with the former approach, the copy is protected
against unwanted changes. In addition, it is not possible to reconstruct
an exception from the standard output.

The new module testing was introduced to provide reusable integration
tests.