Add support for inversed stack traces
When using the stdout to send out logs it is common to truncate the
stacktrace to save space and increase performance. It is also common to
reorder the stacktrace so that the root cause is printed first.

This change introduces support for stack traces where the root cause is
printed first. For this job the custom parser is replaced by an ANTLR4
generated parser. This allows us to create a more flexible processing of
the log output.

Java's stack trace format is awkward to parse because it partly uses
indention to indicated depth in nested elements. Such a format is hard to
parse with a tool like ANTLR. To simplifiy the job the input is
preformatted by a custom reader. The reader adds special opening and
closing sequences so that the parser itself is simpler.