Fix parser errors with Java 9+
Java 9 changed the string representation of stack traces. The
representation might now include the class loader and/or the module name.
The parser did not support that until now. With this change the class
loader/module name will be ignored.

In addition this change fixes a parsing error that could occur when a
incomplete logging line was written to the output.