Improve parsing of standard out
The parser now supports more uncommon outputs e.g. multiple lines,
indentation in log messages, left out line numbers, exotic file, class
or method names. Due to the format that is used by Logstash the output is
not always unambiguous. The parser is designed in such a way that it
interprets the output in

This DetendReader is replaced by an ANTLR based solution. The
DetendContentListener now preprocesses the stack trace for the next parser
step. The DetendContentListener is better then the DetendReader in
distinguishing between messages and types in the stack trace.

This commit also introduces property based testing. The testing is used to
find edge cases in which the parser does not work.