Fix duplicated writing to system out
Output to the system out was partially duplicated due to a defect in the
InterceptablePrintStream. The implementation called super in the
overwritten methods. The super methods called public methods on the same
instance. Because all methods are virtual the delegator
InterceptablePrintStream was called again. E.g. a println caused that
super was called, that itself called print before delagating the call.
The print itself called print on super again before delegating. By this
the print part was duplicated.

PrintStream is a concrete class and acts as a interface at the same time.
For the delegator we only want the interface part. As a solutation for that
the implemenation of PrintStream is fed with a OutputStream that throws
everything away. In addition all methods are overwritten and no super
is called in them. What remains is the interface.

In order to detect such defects earlier, property tests are introduced.
They create random sequences of method calls on InterceptablePrintStream
and check if it acts as an transparent in place replacement for the
PrintStream implementation.