Support spying in more testing frameworks
This change introduces a new set of functions that allows to spy on the
logs in many testing frameworks. For that the dependencies of the core
module from Junit 5 were removed. The Junit 5 functionality was moved
into the new module junit5-support.

The new approach also allowed to split the integration tests for the
Junit extension into those that test the integration in Junit 5 and the
integration of the log spy provider into the logging backend.