Improve provided Hamcrest matchers
The Hamcrest matchers are now alligned so that they can be combined with
existing matchers. For that supplements for the Hamcrest contains/hasItem
matchers were written. The Hamcrest contains/hasItem matchers have a
messed up method signature because of generics.

The matchers are now tested. Also the Jacoco Gradle setup was aligned to
the suggested way in the Gradle documentation.