Improve Java support
Until now it was impossible to create log spies for classes/painful to
create log spies for named loggers from Java. This commit introduces
methods for Java to easily create log spies from Java.

Java can not call Kotlin reified functions. Reified functions are used
to create log spies by type in Kotlin. Furthermore the Kotlin functions
use lambda expression with the return type Unit to pass executable blocks.
Those lambda expressions can be used by Java but provide a bad developer
experience.

The closest Java equivalent would be a Runnable. Kotlin however requires to
use its Function0 interface and returning a value. This adds unnecessary
clutter in Java.

The commit also bumps the version of the formatter to 0.6.4. This is
necessary because the formatter did not the fun keyword for interfaces in
previous version.