Run tests in different Java versions
The result of the integration tests can depend on the Java version in use.
E.g. the output of the stack trace changed over time. In Java 9 the output
was extended by the module name. To detect such specifics the tests will
run in different Java version. All Java versions equal to 8 or greater that
are not end of life are tested.